Add files via upload
Update to support commas in fields. The only types of fields that will not work are fields that begin with open ended commas, such as 
ex: "stuff
or fields that have a quote followed by a comma in the middle of it like the following
ex: "wow", she said
All other fields work with no issues.